refactor: 더 명확한 이름 사용, 코드 다듬기
- useVolumeAnalyser -> useSpeakerHightlighter
