+++ frontend/src/utils/unit8Array.ts
@@ -1,0 +1,3 @@
+export function calculateAverage(array: Uint8Array) {
+  return array.reduce((acc, value) => acc + value, 0);
+}
